added ability to enable/disable system and IP added regexGroups to message templates fixed bug (spring context generated &#xa; instead of \n) fixed tests added mock services for tests added checksum of context to dynamically regenerate it on startup added dispatching mechanism added mainConfig elements to MockConfig.xsd added log rotation mechanism increased version

+++ mock-interactive/src/main/java/ru/sbt/bpm/mock/config/entities/MainConfig.java
@@ -18,7 +18,7 @@ public class MainConfig {
     private String driverTimeout;
 
     @XStreamAlias("maxLogsCount")
-    private Long maxLogsCount;
+    private Integer maxLogsCount;
 
     @XStreamAlias("validationEnabled")
     private Boolean validationEnabled;
